ChatWithDb/requirements.txt
mysql-connector-python>=8.0.33
groq>=0.4.0
python-dotenv>=1.0.0
pandas